system/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
ultralytics>=8.0.0
pillow>=10.0.0
python-multipart==0.0.9
